refactor: add types folder
